requirements.txt
altair==5.4.1
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.7.0
cachetools==5.5.0
category_encoders==2.7.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.10
dash==3.0.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.18.0
Flask==3.0.3
fonttools==4.51.0
fqdn==1.5.1
gitdb==4.0.11
GitPython==3.1.43
gunicorn
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.7
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.5
lightgbm==4.6.0
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
narwhals==1.13.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
orjson==3.10.1
overrides==7.7.0
packaging==24.0
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pillow==10.3.0
platformdirs==4.2.1
plotly==5.21.0
plotly-resampler==0.10.0
pmdarima==2.0.4
prometheus_client==0.21.0
prompt-toolkit==3.0.43
protobuf==5.28.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==18.0.0
pycaret==3.3.2
pycparser==2.22
pydeck==0.9.1
Pygments==2.17.2
pyod==2.0.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-whois==0.9.5
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.0
requests==2.31.0
requests-file==2.1.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.18.0
schemdraw==0.15
scikit-base==0.7.7
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
Send2Trash==1.8.3
six==1.16.0
sklearn-compat==0.1.3
sktime==0.26.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.39.0
tbats==1.1.3
tenacity==8.2.3
terminado==0.18.1
threadpoolctl==3.4.0
tinycss2==1.4.0
tldextract==5.1.3
toml==0.10.2
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
tsdownsample==0.1.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
watchdog==5.0.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.2
whois==1.20240129.2
widgetsnbextension==4.0.13
xgboost==2.1.2
xxhash==3.4.1
yellowbrick==1.5
zipp==3.21.0
